Add bold styling to certain words and phrases
I used the <b> tag to bold some words where only the bold styling was necessary, not the higher importance. In that section those words need to be bold only visually.

I used the <strong> tag to some phrases to indicate their higher importance with the rest of the list item.